+++ template/src/templates/httpbin.rs
@@ -4,28 +4,32 @@ impl LapsePreset {
   pub fn httpbin() -> Self {
     use TemplateEntry::File;
 
-    Self::new(vec![
+    let base = Self::default();
+
+    let superset = Self::new(vec![
       File(
-        "get.md".to_string(),
+        "requests/get.md".to_string(),
         include_str!("../../templates/httpbin/requests/get.md").to_string(),
       ),
       File(
-        "post.md".to_string(),
+        "requests/post.md".to_string(),
         include_str!("../../templates/httpbin/requests/post.md").to_string(),
       ),
       File(
-        "patch.md".to_string(),
+        "requests/patch.md".to_string(),
         include_str!("../../templates/httpbin/requests/patch.md").to_string(),
       ),
       File(
-        "delete.md".to_string(),
+        "requests/delete.md".to_string(),
         include_str!("../../templates/httpbin/requests/delete.md").to_string(),
       ),
       File(
-        "put.md".to_string(),
+        "requests/put.md".to_string(),
         include_str!("../../templates/httpbin/requests/put.md").to_string(),
       ),
-    ])
+    ]);
+
+    base + superset
   }
 }