+++ CHANGELOG.md
@@ -17,6 +17,7 @@ https://github.com/gab1618/lapse/compare/latest..HEAD
 - Timestamp and duration informations to log
 - Response showing pretty response log
 - Default scheme configurable by env
+- Env config file
 
 ### Fixed